sua code theo review cua anh Khoi
