+++ src/App.tsx
@@ -8,7 +8,7 @@ interface route {
 }
 function App() {
     const showContentMenus = (routes: route[]) => {
-        var result = null;
+        let result = null;
         if (routes.length > 0) {
             result = routes.map((route, index) => {
                 return (
