+++ src/index.tsx
@@ -19,4 +19,5 @@ ReactDOM.render(
     </Provider>,
     document.getElementById('root')
 );
+export default store;
 //registerServiceWorker();
