+++ src/pages/NotFound/NotFound.tsx
@@ -1,9 +1,6 @@
-import React from 'react';
-
 const NotFound = () => {
     return (
         <h1>Không Tìm Thấy Trang</h1>
     );
 }
-
 export default NotFound;
