+++ src/utils/apiCaller.tsx
@@ -1,23 +1,7 @@
 import axios from 'axios';
 import * as Config from '../constants/Config';
-export type Method =
-    | 'get' | 'GET'
-    | 'delete' | 'DELETE'
-    | 'head' | 'HEAD'
-    | 'options' | 'OPTIONS'
-    | 'post' | 'POST'
-    | 'put' | 'PUT'
-    | 'patch' | 'PATCH'
-    | 'purge' | 'PURGE'
-    | 'link' | 'LINK'
-    | 'unlink' | 'UNLINK'
-type product = {
-    id?: string,
-    name: string,
-    description: string,
-    price: number,
-    status: boolean
-}
+import type { product, Method } from '../constants/Types';
+
 export default function callApi(endpoint: string, method: Method, data: product | null) {
     return axios({
         method: method,
